requirements.txt
PyQt5
requests
beautifulsoup4
python-docx
fuzzywuzzy
python-Levenshtein
lxml 